simplify timer class
git-svn-id: https://obdcon.svn.sourceforge.net/svnroot/obdcon@16 6dd02fe3-967f-4638-bac9-8d06f6e08789

+++ libctb/iobase.h
@@ -211,7 +211,7 @@ namespace ctb {
 	   virtual int ReadUntilEOS(char*& readbuf,
 						   size_t* readedBytes,
 						   char* eosString = "\n",
-						   long timeout_in_ms = 1000L,
+						   unsigned int timeout_in_ms = 1000L,
 						   char quota = 0);
 
 	   /*!
